[svn-r11217] Purpose:     Bug fix & code cleanup
Description:
    Fix another bug in the file mounting code and refactor the unmount
code that it is simpler.

Platforms tested:
    FreeBSD 4.11 (sleipnir)
    Linux 2.4
    Too minor to require h5committest

+++ tools/h5diff/h5diff_main.c
@@ -157,7 +157,7 @@ int main(int argc, const char *argv[])
        printf("<-n %s> is not a valid option\n", argv[i+1] );
        usage();
       }
-      options.count = atoi(argv[i+1]);
+      options.count = atol(argv[i+1]);
       i++; /* go to next */
      }
      else
